queue: Add first body of code
This queue, whilst not lock free, is designed to be non-blocking.

This means that locks are tried and when found not to be successful,
their neighbours will be tried, or a scan will take place in order to
free up some locks.

There are only a few cases in which a query will fail.

The first of which, the queue being empty and the buffer being empty,
whilst trying to fetch data from it.

The second, trying to write data to the buffer, but the buffer being
full, whilest another thread is flushing the buffer.

The third, trying to read from the queue while a flush is in progress
and there is only a single object in queue.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>